+++ src/main/java/com/sudo/railo/train/application/SeatService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import com.sudo.railo.train.config.TrainTemplateProperties;
 import com.sudo.railo.train.domain.Seat;
 import com.sudo.railo.train.domain.TrainCar;
-import com.sudo.railo.train.infrastructure.TrainJdbcRepository;
+import com.sudo.railo.train.infrastructure.jdbc.TrainJdbcRepository;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
